PDO-4847 - Add weekly pipeline run logic for PCB

See merge request ping-cloud-private-tenant/ping-cloud-base!1879

+++ tests/chaos/01-delete-pa-admin-pod.sh
@@ -12,6 +12,11 @@ fi
 
 testDeletePaAdmin() {
 
+  if [ "${ENV_TYPE}" == "customer-hub" ]; then
+    log "Customer-hub deployment, skipping test"
+    return 0
+  fi
+
   PA_ADMIN_PASSWORD=${PA_ADMIN_PASSWORD:-2FederateM0re}
 
   kubectl delete pod pingaccess-admin-0 -n "${PING_CLOUD_NAMESPACE}"
